restructured application for multipage content
