+++ Morocco dashboard/scripts/plotting.py
